PROJETS/PRO/SUBLYM/requirements.txt
# SUBLYM — Dépendances

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Image generation
replicate>=0.20.0

# Future: video generation
# minimax-sdk
